+++ cocos/core/pipeline/render-batched-queue.ts
@@ -8,7 +8,18 @@ import { BatchedBuffer } from './batched-buffer';
 import { PipelineStateManager } from './pipeline-state-manager';
 import { GFXDevice } from '../gfx/device';
 import { GFXRenderPass } from '../gfx';
-import { Light } from '../renderer/scene/light';
+import { IPSOCreateInfo } from '../renderer';
+import { IRenderObject, UBOForwardLight } from './define';
+import { LightType, Light } from '../renderer/scene/light';
+import { IMacroPatch, Pass } from '../renderer/core/pass';
+
+const spherePatches: IMacroPatch[] = [
+    { name: 'CC_FORWARD_ADD', value: true },
+];
+const spotPatches: IMacroPatch[] = [
+    { name: 'CC_FORWARD_ADD', value: true },
+    { name: 'CC_SPOTLIGHT', value: true },
+];
 
 /**
  * @en The render queue for dynamic batching
@@ -22,6 +33,35 @@ export class RenderBatchedQueue {
      */
     public queue = new Set<BatchedBuffer>();
 
+    private static _lightPsoCreateInfos: IPSOCreateInfo[] = [];
+
+    public static getLightPipelineCreateInfo (renderObj: IRenderObject, subModelIdx: number, pass: Pass,
+        validLights: Light[], lightGFXBuffers: GFXBuffer[], lightIdx: number): IPSOCreateInfo {
+        if (!this._lightPsoCreateInfos[lightIdx]) {
+            const modelPatches = renderObj.model.getMacroPatches(subModelIdx);
+            const light = validLights[lightIdx];
+            const lightBuffer = lightGFXBuffers[lightIdx];
+
+            let fullPatches: IMacroPatch[] = [];
+            switch (light.type) {
+                case LightType.SPHERE:
+                    fullPatches = modelPatches ? spherePatches.concat(modelPatches) : spherePatches;
+                    break;
+                case LightType.SPOT:
+                    fullPatches = modelPatches ? spotPatches.concat(modelPatches) : spotPatches;
+                    break;
+            }
+
+            const psoci = pass.createPipelineStateCI(fullPatches)!;
+            this._lightPsoCreateInfos[lightIdx] = psoci;
+            renderObj.model.updateLocalBindings(psoci, subModelIdx);
+            psoci.bindingLayout.bindBuffer(UBOForwardLight.BLOCK.binding, lightBuffer);
+            psoci.bindingLayout.update();
+        }
+
+        return this._lightPsoCreateInfos[lightIdx];
+    }
+
     /**
      * @en Clear the render queue
      * @zh 清空渲染队列。
@@ -30,19 +70,6 @@ export class RenderBatchedQueue {
         const it = this.queue.values(); let res = it.next();
         while (!res.done) {
             res.value.clear();
-            res = it.next();
-        }
-        this.queue.clear();
-    }
-
-    /**
-     * @en Clear the render queue
-     * @zh 清空渲染队列。
-     */
-    public clearLightBatched (validLights: Light[], lightBuffers: GFXBuffer[]) {
-        const it = this.queue.values(); let res = it.next();
-        while (!res.done) {
-            res.value.clearLightBatched(validLights, lightBuffers);
             res = it.next();
         }
         this.queue.clear();
